refactor: :art: add config for absolute import
