+++ components/footer/index.js
@@ -1,6 +1,6 @@
 import { Center, HStack, Icon, Link, Text, VStack } from '@chakra-ui/react'
-import { github, linkedin, twitter } from '../../data/socials'
-import useToggle from '../../utils/hooks/useToggle'
+import { github, linkedin, twitter } from '@/data/socials'
+import useToggle from '@/hooks/useToggle'
 
 const Footer = () => (
 	<VStack
