+++ components/styled/index.js
@@ -1,6 +1,6 @@
-import Label from './label'
-import StyledLink from './link'
+import Label from '@/components/styled/label'
+import StyledLink from '@/components/styled/link'
 
 export { StyledLink }
 export { Label }
-export { SuccessMessage, ErrorMessage } from './message'
+export { SuccessMessage, ErrorMessage } from '@/components/styled/message'
