+++ jsconfig.json
@@ -1,0 +1,14 @@
+{
+	"compilerOptions": {
+		"baseUrl": ".",
+		"paths": {
+			"@/components/*": ["components/*"],
+			"@/data/*": ["data/*"],
+			"@/layouts/*": ["layouts/*"],
+			"@/lib/*": ["lib/*"],
+			"@/styles/*": ["styles/*"],
+			"@/utils/*": ["utils/*"],
+			"@/hooks/*": ["utils/hooks/*"]
+		}
+	}
+}
