+++ pages/projects.js
@@ -1,8 +1,8 @@
 import { Box, Grid, Heading, Text } from '@chakra-ui/react'
-import Container from '../layout/container'
-import ContentWrapper from '../layout/contentWrapper'
-import projects from '../data/projects'
-import ProjectCard from '../components/projectCard'
+import Container from '@/layouts/container'
+import ContentWrapper from '@/layouts/contentWrapper'
+import projects from '@/data/projects'
+import ProjectCard from '@/components/projectCard'
 
 const Projects = () => (
 	<Container title='Projects | Nitesh Seram'>
